# docs/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sub2API
  description: |
    轻量级 API 聚合中转服务，提供 OpenAI 兼容格式的统一接口。
    
    ## 认证
    所有 API 请求需要在 Header 中携带 API Key：
    ```
    Authorization: Bearer sk-sub2api-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```
    
    ## 错误处理
    错误响应遵循统一格式：
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "Human readable message",
        "type": "error_type"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: Sub2API Support
servers:
  - url: https://api.sub2api.com
    description: Production
  - url: http://localhost:3000
    description: Local Development
...
paths:
...
  /v1/keys:
    get:
      tags: [Keys]
      summary: List user's API keys
      description: 返回当前用户的所有 API Key（不含完整密钥）
      operationId: listKeys
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Key list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Keys]
      summary: Create new API key
      description: |
        创建新的 API Key。
        
        ⚠️ 完整的 Key 值仅在此响应中返回一次，之后无法再获取。
      operationId: createKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
      responses:
        '201':
          description: Key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyWithSecret'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/keys/{key_id}:
    get:
      tags: [Keys]
      summary: Get key details
      operationId: getKey
      security:
        - BearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Key details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Keys]
      summary: Update key
      description: 更新 Key 的名称或额度
      operationId: updateKey
      security:
        - BearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateKeyRequest'
      responses:
        '200':
          description: Key updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Keys]
      summary: Disable key
      description: 禁用 API Key（软删除）
      operationId: deleteKey
      security:
        - BearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Key disabled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Usage & Billing
  # ============================================
  /v1/usage:
    get:
      tags: [Usage]
      summary: Get usage summary
      description: 返回余额和用量统计
      operationId: getUsage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/transactions:
    get:
      tags: [Usage]
      summary: List transactions
      description: 返回交易流水记录
      operationId: listTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [topup, consume, refund]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # Admin Operations
  # ============================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        用户 API Key，格式：`sk-sub2api-{32 chars}`
    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: 管理员密钥

  # ============================================
  # Schemas
  # ============================================
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          example: "key_abc123"
        key_prefix:
          type: string
          description: Key 前缀（用于识别）
          example: "sk-sub2api-aBcD..."
        name:
          type: string
          example: "My App Key"
        quota:
          type: number
          nullable: true
          description: 额度限制（null=无限）
        used_amount:
          type: number
          description: 已使用额度
        status:
          type: string
          enum: [active, disabled]
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ApiKeyWithSecret:
      allOf:
        - $ref: '#/components/schemas/ApiKey'
        - type: object
          properties:
            key:
              type: string
              description: 完整 Key（仅返回一次）
              example: "sk-sub2api-aBcDeFgHiJkLmNoPqRsTuVwXyZ012345"
    ApiKeyList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'
    CreateKeyRequest:
      type: object
      properties:
        name:
          type: string
          example: "My App Key"
        quota:
          type: number
          description: 额度限制
        expires_at:
          type: string
          format: date-time
    UpdateKeyRequest:
      type: object
      properties:
        name:
          type: string
        quota:
          type: number

    # --- Usage & Transactions ---
    UsageSummary:
      type: object
      properties:
        balance:
          type: number
          description: 当前余额 (USD)
          example: 9.876543
        currency:
          type: string
          enum: [USD]
        usage_today:
          $ref: '#/components/schemas/UsageStats'
        usage_month:
          $ref: '#/components/schemas/UsageStats'
    UsageStats:
      type: object
      properties:
        requests:
          type: integer
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost:
          type: number
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [topup, consume, refund]
        amount:
          type: number
          description: 金额（consume 为负）
        balance_after:
          type: number
        model:
          type: string
          nullable: true
        input_tokens:
          type: integer
          nullable: true
        output_tokens:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
    TransactionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # --- Admin ---
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            type:
              type: string

  # ============================================
  # Responses
  # ============================================
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "invalid_api_key"
              message: "Invalid API key"
              type: "authentication_error"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "not_found"
              message: "Resource not found"
              type: "not_found_error"
...